chore: refactor build output structure
