+++ tsconfig.json
@@ -3,6 +3,7 @@
     "outDir": "dist",
     "target": "ES2021",
     "module": "node16",
+    "rootDir": "src",
     "moduleResolution": "node16",
     "declaration": false,
     "isolatedModules": true,
@@ -21,5 +22,6 @@
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
-  }
+  },
+  "include": ["src/**/*"]
 }
